Bug fiex in plotPosts.R
Since bin widths are calculated according to the SD of the data, when the data is invariant, either becuase the parameter isn't estimated, it's not an actual parameter (e.g. Tree.no) nor because signal is lost deep in the tree (for example root nodes), there is an SD of zero, and thus a bin width of zero. This is obviously not OK, and so now there is an IF statement that changes the binwidth to 1 if it is 0. This just makes it easier to plot all the posteriors (including non-parameters) without errors. This is the fastest way to see all posteriors, and you can just ignore the non-parameters.